fixing Branch() by generalizing it
In trying to clean up the Branch() constructor  (i.e. remove the internal state variable), I realized that what it was originally attempting to accomplish was the ability to "map out" branches (heh) of arbitrary lexing behavior (representable with Acceptors) on a per rune (one or more) basis many runes in advance. That is, decide on the Acceptor for the next rune based on logic in the Acceptor for the current rune being lexed. I arrived at this by realizing that an easy solution would be possible should the anonymous Acceptor returned by Branch() have access to the variable to which it was assigned. The logic in the current Acceptor could then /directly/ change the next Acceptor by overwriting itself wherever it was being used. After all, the first solution I tried seemed to be attempting just that by having another nested layer with all that internal state hubbub.

If I squint hard enough, I see a monad, so the next step will be to explore the application of monadic ideas to Branch() or more generally the sequencing of lexer behavior as modelled by Acceptors. Functions are maps are functions are maps, etc as it is -- perhaps the map data structure can be avoided altogether.

Interesting tidbit: In Go, anonymous functions can't recurse without having access to their external context! https://github.com/golang/go/issues/226'